file not visible in SPIFFS_readdir #90
Iterator bug - if an iteration was callbacked at last page in block -1,
the resumed iteration would skip the last page in block.
Fixing this revealed another bug, where free page marker was not advanced
but pointing to the same page found before when searching for new free page.
This is probably not a problem, as when searching for free pages the found
free page always becomes occupied. A new free page scan will simply skip
the previously found free page, which has been allocated by now. However,
it is suboptimal. Also, it might have messed with the wear leveling a bit.